Fix type cast

+++ OTAnalytics/plugin_datastore/track_geometry_store/shapely_store.py
@@ -95,7 +95,7 @@ class InvalidTrackGeometryDataset(Exception):
 
 def distance_on_track(point: Point, track_geom: GeometryCollection) -> float:
     distance = line_locate_point(track_geom, point)
-    return round(distance, NDIGITS_DISTANCE)
+    return float(round(distance, NDIGITS_DISTANCE))
 
 
 class ShapelyTrackGeometryDataset(TrackGeometryDataset):
